GameController, now has game.turn implemented. So each time the game goes to next phase it'll update stuff
gameTurnAction() in gameController, makes host update the game state, at each phase (gameTurn).  The game turn only changes, when all the players are Ready other than in turn 0 and turn 5. Turn 5 has some code in it but it's not finished, but rather a template to work on more later. Turn 0,1,2,3,4 work mostly as intended for changing the Turn, and updating the game state

ALSO use this.props.(games, myCards or whatever) in GameSettings.js. "this.props" is the exact passed in property from GameController, and the only way to use the autoupdated version of the state.

turn = 0 //initial, shuffle cards and upload to database,
  turn = 1 //buy in phase and distrute cards to players
  turn = 2 //place 3 cards on board, and players can fold/raise/check/call
  turn = 3 //place 4th card, bet
  turn = 4 //place 5th card, bet
  turn = 5 //show cards, last turn and winner takes pot. RESET turn to 0