LeaveGame works solo, userData prop, Games now Cost Chips
LeaveGame works solo. If only 1 user is in game and leaves the game will be deleted, and all chips, and stats will be updated for player leaving, such as how many chips won, and their new balance.

Create game now subtracts the Buy In amount from your balance. If you can't afford your own buy in an Alert shows telling you to lower the Buy-In. Also if you are currently in a game you cannot create a new game, and a different alert is shown for that.

There is now a userData prop that holds all the user's information. This prop is passed into some classes that need it, and is created in App.js. Feel free to use this prop where ever needed, instead of calling from Firebase all the time.

I also Fixed Change Email, and Change Username to work as intended. Balance now works, as long as your Signed In, and will always be up to date unlike previously.

Refined gameTurnAction() in GameController to work better.

LandingPage has "Continue Game" to join game in progress, and "Join Game (TBA)" that just goes to Create Game for now. Depending on if you are in a game or not either one will show.