Test cases for Game Controller Functions Written
Need to install Jest if you wish to run.
Written but not actually tested/ran test cases for some functions in Game Controller such as card/hand rankings. I am not sure how to exactly test the functions if some data is from the database, so I wrote them as if they were.