Raise/All in improved and Auto Kick implemented
All players can only raise up to the amount which the player with lowest balance can go All In!

All in occurs when someone raises to the full amount of the lowest player balance, whether it's that player them self, or another player

Player is automatically kicked at game.turn == 0 when their balance is below the blindAmount.